Fix model
